services/excel_filler/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
openpyxl==3.1.5
PyYAML==6.0.2
pydantic==2.10.6
python-multipart==0.0.20
pytest==8.3.4
httpx==0.28.1
